move methods to lifecycle where they belong
